Use full class names in specs
Originally I thought it was worth shortening the class names in the
specs but now I think it makes it less readable since it 1) doesn't
syntax highlight and 2) requires looking at the top of the spec to know
what `ok` and `err` refer to.

I considered aliasing the constants like I suggest in the README but
that would pollute the top-level namespace in the other specs which
seems undesirable.